+++ pegainfer-qwen35/src/scheduler.rs
@@ -7,6 +7,7 @@
 mod backend;
 mod plan;
 mod steps;
+mod telemetry;
 mod tp;
 use std::collections::HashMap;
 use std::collections::HashSet;
@@ -59,6 +60,7 @@ use self::plan::plan_prefill_chunks;
 use self::plan::prefilling_future_pages;
 use self::plan::slot_for_new_request;
 use self::steps::*;
+use self::telemetry::*;
 use self::tp::*;
 use crate::Qwen35DecodeOverlap;
 use crate::Qwen35SchedulerPolicy;
@@ -288,43 +290,6 @@ pub const DEFAULT_MAX_PREFILL_TOKENS: usize = 1024;
 /// decode, completion, and wait actions instead of relying on the coarse
 /// `[submit, last-token]` injection window. Off by default: no cost on the
 /// normal bench path.
-fn itl_debug_enabled() -> bool {
-    static ENABLED: OnceLock<bool> = OnceLock::new();
-    *ENABLED.get_or_init(|| std::env::var_os("PEGAINFER_ITL_DEBUG").is_some())
-}
-
-/// Monotonic microseconds since the first ITL step, so `ITL_STEP` timestamps
-/// are correlatable within one process run (paired with wall-clock epoch us).
-fn itl_debug_mono_us() -> u128 {
-    static ORIGIN: OnceLock<Instant> = OnceLock::new();
-    ORIGIN.get_or_init(Instant::now).elapsed().as_micros()
-}
-
-fn log_itl_step(
-    step_start: Option<Instant>,
-    plan: &str,
-    prefill_tokens: usize,
-    prefill_reqs: usize,
-    decode_n: usize,
-) {
-    let Some(step_start) = step_start else {
-        return;
-    };
-    let dur_us = step_start.elapsed().as_micros();
-    let epoch_us = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map_or(0, |d| d.as_micros());
-    info!(
-        "ITL_STEP mono_us={} epoch_us={} plan={} prefill_tok={} prefill_reqs={} decode_n={} dur_us={}",
-        itl_debug_mono_us(),
-        epoch_us,
-        plan,
-        prefill_tokens,
-        prefill_reqs,
-        decode_n,
-        dur_us
-    );
-}
 
 // ── Entry point ─────────────────────────────────────────────────────────
